+++ src/mkvtoolnix-gui/merge/track.h
@@ -95,6 +95,7 @@ protected:
   void setDefaultsForcedDisplayFlag();
   void setDefaultsCommentaryFlag();
   void setDefaultsHearingImpairedFlag();
+  void setDefaultsOriginalLanguageFlag();
   void setDefaultsDisplayDimensions();
   void setDefaultsLanguage(mtx::bcp47::language_c const &languageDerivedFromFileName);
   void setDefaultsColor();